+++ crates/blockchain/blockchain.rs
@@ -190,7 +190,7 @@ impl Blockchain {
             AccountUpdatesList,
             // FIXME: extract to stats struct
             usize,
-            [Instant; 7],
+            [Instant; 6],
         ),
         ChainError,
     > {
@@ -217,10 +217,16 @@ impl Blockchain {
         let mut max_queue_length = 0;
         let (execution_result, account_updates_list) = std::thread::scope(|s| {
             let (tx, rx) = channel();
-            let execution_handle = s.spawn(move || {
-                let execution_result = vm.execute_block_pipeline(block, tx, queue_length_ref);
+            let execution_handle = s.spawn(move || -> Result<_, ChainError> {
+                let execution_result = vm.execute_block_pipeline(block, tx, queue_length_ref)?;
+
+                // Validate execution went alright
+                validate_gas_used(&execution_result.receipts, &block.header)?;
+                validate_receipts_root(&block.header, &execution_result.receipts)?;
+                validate_requests_hash(&block.header, &chain_config, &execution_result.requests)?;
+
                 let exec_end_instant = Instant::now();
-                execution_result.map(move |r| (r, exec_end_instant))
+                Ok((execution_result, exec_end_instant))
             });
             let merkleize_handle = s.spawn(move || -> Result<_, StoreError> {
                 let account_updates_list = self.handle_merkleization(
@@ -234,7 +240,7 @@ impl Blockchain {
             });
             (
                 execution_handle.join().unwrap_or_else(|_| {
-                    Err(EvmError::Custom("execution thread panicked".to_string()))
+                    Err(ChainError::Custom("execution thread panicked".to_string()))
                 }),
                 merkleize_handle.join().unwrap_or_else(|_| {
                     Err(StoreError::Custom(
@@ -247,12 +253,6 @@ impl Blockchain {
         let (execution_result, exec_end_instant) = execution_result?;
         let exec_merkle_end_instant = Instant::now();
 
-        // Validate execution went alright
-        validate_gas_used(&execution_result.receipts, &block.header)?;
-        validate_receipts_root(&block.header, &execution_result.receipts)?;
-        validate_requests_hash(&block.header, &chain_config, &execution_result.requests)?;
-        let results_validated_instant = Instant::now();
-
         Ok((
             execution_result,
             account_updates_list,
@@ -264,7 +264,6 @@ impl Blockchain {
                 exec_end_instant,
                 merkle_end_instant,
                 exec_merkle_end_instant,
-                results_validated_instant,
             ],
         ))
     }
@@ -942,9 +941,8 @@ impl Blockchain {
             exec_end_instant,
             merkle_end_instant,
             exec_merkle_end_instant,
-            results_validated_instant,
             stored_instant,
-        ]: [Instant; 8],
+        ]: [Instant; 7],
     ) {
         let interval = stored_instant.duration_since(start_instant).as_secs_f64();
         if interval != 0f64 {
@@ -974,13 +972,12 @@ impl Blockchain {
             };
             let extra_log = if as_gigas > 0.0 {
                 format!(
-                    " block validation: {}% exec+merkle: {}% (exec: {}% merkle: {}% max_queue_length: {merkle_queue_length}) results validation: {}% store: {}%",
+                    " block validation: {}% exec+merkle: {}% (exec: {}% merkle: {}% max_queue_length: {merkle_queue_length}) store: {}%",
                     percentage(start_instant, block_validated_instant),
                     percentage(exec_merkle_start, exec_merkle_end_instant),
                     percentage(exec_merkle_start, exec_end_instant),
                     percentage(exec_merkle_start, merkle_end_instant),
-                    percentage(merkle_end_instant, results_validated_instant),
-                    percentage(results_validated_instant, stored_instant),
+                    percentage(merkle_end_instant, stored_instant),
                 )
             } else {
                 "".to_string()