+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Perf
 
+### 2025-11-03
+- Merge execution with some post-execution validations [#5170](https://github.com/lambdaclass/ethrex/pull/5170)
+
 ### 2025-10-31
 - Reduce overhead of trie opening [#5145](https://github.com/lambdaclass/ethrex/pull/5145)
 - Improved discovery and peer initialization [#5147](https://github.com/lambdaclass/ethrex/pull/5147)